requirements.txt
asgiref==3.8.1
colorama==0.4.6
Django==5.0.7
factory_boy==3.3.1
Faker==27.0.0
iniconfig==2.0.0
packaging==24.1
pillow==10.4.0
pluggy==1.5.0
pytest==8.3.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
whitenoise==6.7.0
